apps/idp-api/pyproject.toml
[project]
name = "idp-api"
version = "0.1.0"
description = "GitOps IDP backend API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "pydantic>=2.10.0",
  "pyyaml>=6.0.0"
]

[tool.uv]
dev-dependencies = []
